* CC: fix a crash issue (buffer overflow) introduced in rev11545.
In rev11545, I changed the way of checking the end of buffer. Before r11545,
an ending \0 char is checked to see whether we reaches the end of buffer.
After r11545, I use the KMP_Find(p, bufferLen, ...), but here the p will
increase when we hit a match, while the bufferLen does not decrease.

In this commit, when p increased, the searchLength is decreased by the same
value.